pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fakephylo"
version = "0.1.0"
description = "Tutorial on building phylogenetic trees using UPGMA with flexible distance metrics"
requires-python = ">=3.8"
dependencies = [
    "biopython>=1.81",
    "numpy>=1.24.0",
]

[tool.setuptools.packages.find]
include = ["fakephylo*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "python-Levenshtein>=0.21.0",
    "mkdocs-material>=9.0.0",
    "mypy>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
